requirements.txt
pandas>=1.5.0 
numpy>=1.23.0 
fastapi>=0.104.0 
uvicorn>=0.23.0 
transformers>=4.30.0 
torch>=2.0.0 
spacy>=3.5.0 
fuzzywuzzy>=0.18.0 
scipy>=1.9.0 
pyyaml>=6.0.0 
python-multipart>=0.0.6